+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/post-unity-activate-license.ts
@@ -11,6 +11,7 @@ import {
     localPathInputVariableName,
     passwordInputVariableName,
     unityEditorsPathModeInputVariableName,
+    unityProjectPathInputVariableName,
     usernameInputVariableName,
     versionInputVariableName,
     versionSelectionModeVariableName
@@ -32,6 +33,7 @@ function run() {
         const unityEditorsPath = UnityPathTools.getUnityEditorsPath(
             tl.getInput(unityEditorsPathModeInputVariableName, true)!,
             tl.getInput(customUnityEditorsPathInputVariableName));
+        const unityProjectPath = tl.getPathInput(unityProjectPathInputVariableName) || '';
 
         var unityVersion: UnityVersionInfoResult;
         if (versionSelectionMode === 'specify') {
@@ -61,6 +63,7 @@ function run() {
             const unityCmd = tl.tool(unityExecutablePath)
                 .arg('-batchmode')
                 .arg('-nographics')
+                .arg('-projectPath').arg(unityProjectPath)
                 .arg('-username').arg(username)
                 .arg('-password').arg(password)
                 .arg('-returnlicense')